# specs/speckit-plus-phase-ii/contracts/api-contract.yaml
openapi: 3.0.3
info:
  title: Speckit Plus Phase II - Todo Application API
  description: API for managing user tasks in the Speckit Plus todo application
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.speckitplus.example.com
    description: Production server
paths:
  /api/{user_id}/tasks:
    get:
      summary: Retrieve all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: search
          in: query
          schema:
            type: string
          description: Search keywords in title and description
        - name: priority
          in: query
          schema:
            type: string
            enum: [Critical, High, Medium, Low]
          description: Filter by priority level
        - name: status
          in: query
          schema:
            type: string
            enum: [complete, incomplete]
          description: Filter by completion status
        - name: timestamp_from
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by timestamp range (from)
        - name: timestamp_to
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by timestamp range (to)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
      security:
        - bearerAuth: []
    post:
      summary: Create a new task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{id}:
    get:
      summary: Retrieve a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
      security:
        - bearerAuth: []
    put:
      summary: Update a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
      security:
        - bearerAuth: []
    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                complete:
                  type: boolean
                  description: Whether the task is complete
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
      security:
        - bearerAuth: []

  /api/auth/signup:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
        '409':
          description: Conflict - email already exists
  /api/auth/signin:
    post:
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /api/auth/signout:
    post:
      summary: Logout user
      responses:
        '200':
          description: Logout successful
      security:
        - bearerAuth: []
  /api/auth/password:
    put:
      summary: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassword'
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []
  /api/auth/forgot-password:
    post:
      summary: Recover password using security question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPassword'
      responses:
        '200':
          description: Password recovery initiated
        '400':
          description: Bad request
        '404':
          description: User not found
  /api/auth/account:
    delete:
      summary: Delete user account
      responses:
        '204':
          description: Account deleted successfully
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - priority
        - timestamp
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns the task
        title:
          type: string
          maxLength: 255
          description: Title of the task
        description:
          type: string
          maxLength: 1000
          description: Optional description of the task
        priority:
          type: string
          enum: [Critical, High, Medium, Low]
          description: Priority level of the task
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the task was created or scheduled
        status:
          type: boolean
          default: false
          description: Whether the task is completed
        created_at:
          type: string
          format: date-time
          description: When the task was created
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated

    CreateTaskRequest:
      type: object
      required:
        - title
        - priority
      properties:
        title:
          type: string
          maxLength: 255
          description: Title of the task (required)
        description:
          type: string
          maxLength: 1000
          description: Optional description of the task
        priority:
          type: string
          enum: [Critical, High, Medium, Low]
          description: Priority level of the task
        timestamp:
          type: string
          format: date-time
          description: Timestamp for the task (defaults to current time)

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          description: Updated title of the task
        description:
          type: string
          maxLength: 1000
          description: Updated description of the task
        priority:
          type: string
          enum: [Critical, High, Medium, Low]
          description: Updated priority level of the task
        status:
          type: boolean
          description: Updated completion status

    User:
      type: object
      required:
        - id
        - first_name
        - last_name
        - email
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
        first_name:
          type: string
          maxLength: 100
          description: User's first name
        last_name:
          type: string
          maxLength: 100
          description: User's last name
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address
        created_at:
          type: string
          format: date-time
          description: When the user account was created

    UserRegistration:
      type: object
      required:
        - first_name
        - last_name
        - email
        - password
        - favorite_teacher
      properties:
        first_name:
          type: string
          maxLength: 100
          description: User's first name
        last_name:
          type: string
          maxLength: 100
          description: User's last name
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address
        password:
          type: string
          minLength: 8
          description: User's password with strength requirements
        favorite_teacher:
          type: string
          description: User's favorite teacher for security question

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: User's password

    ChangePassword:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
          description: Current password
        new_password:
          type: string
          minLength: 8
          description: New password with strength requirements

    ForgotPassword:
      type: object
      required:
        - email
        - favorite_teacher_answer
      properties:
        email:
          type: string
          format: email
          description: User's email address
        favorite_teacher_answer:
          type: string
          description: Answer to the favorite teacher security question

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          default: bearer
          description: Token type
        user:
          $ref: '#/components/schemas/User'